新增sql字段

+++ message-handler/src/main/java/com/vv/handler/handler/impl/EmailHandler.java
@@ -10,7 +10,6 @@ import com.vv.common.dto.model.EmailContentModel;
 import com.vv.common.enums.ChannelType;
 import com.vv.handler.handler.BaseHandler;
 import com.vv.handler.handler.Handler;
-import com.vv.support.domain.ChannelAccount;
 import com.vv.support.domain.MessageTemplate;
 import com.vv.support.utils.AccountUtils;
 import com.vv.support.utils.FileUtils;
@@ -62,7 +61,7 @@ public class EmailHandler extends BaseHandler implements Handler{
      *
      * @return
      */
-    private MailAccount getAccountConfig(Integer sendAccount) {
+    private MailAccount getAccountConfig(Long sendAccount) {
         MailAccount account = accountUtils.getChannelAccountById(sendAccount, MailAccount.class);
         try {
             MailSSLSocketFactory sf = new MailSSLSocketFactory();
